+++ BabyBack/src/main/java/com/example/demo/dto/ProductDto.java
@@ -2,4 +2,4 @@ package com.example.demo.dto;
 
 import java.math.BigDecimal;
 
-public record ProductDto(Long id, String name, String imageUrl, String description, BigDecimal price) {}
+public record ProductDto(Integer id, String name, String imageUrl, String description, BigDecimal price) {}
