+++ BabyBack/src/main/java/com/example/demo/repository/PurchaseOrderRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.PurchaseOrder;
 
-public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, Long> {
+public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, Integer> {
 
 	long countByOrderDate(LocalDate now);
