+++ BabyBack/src/main/java/com/example/demo/repository/SupplierProductRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.Query;
 import com.example.demo.model.Product;
 import com.example.demo.model.SupplierProduct;
 
-public interface SupplierProductRepository extends JpaRepository<SupplierProduct, Long> {
-	List<SupplierProduct> findBySupplierId(Long supplierId);
+public interface SupplierProductRepository extends JpaRepository<SupplierProduct, Integer> {
+	List<SupplierProduct> findBySupplierId(Integer supplierId);
 	
 
 }
